removed explict implementation; added do() to linear model; exposed in CausalModel
